\name{height_data}
\alias{height_data}
\docType{data}
\title{
	height data 
}
\description{
	The heights of 532 girls at birth, four, eight, and twelve months, and three, four, and seven years of age.
}
\usage{
	data(height_data)
}
\format{
	A matrix with 532 observations on 7 variables (height at each age).
}
\references{
	Lopez-Pintado, S. and McKeague, I. W. (2013).
	\emph{Recovering gradients from sparsely observed functional
	  data.} Biometrics 69, 396-404 (2013).
	http://www.columbia.edu/~im2131/ps/growthrate-package-reference.pdf
}
\examples{
	data(height_data)
}
\keyword{datasets}
