\name{growthrate-package}
\alias{growthrate-package}
\docType{package}
\title{
	Recovery of gradients from sparsely observed functional data
}
\description{
	A nonparametric empirical Bayes method for recovering gradients (or growth velocities) from observations of smooth functions (e.g., growth curves) at isolated time points.
}
\references{
	Lopez-Pintado, S. and McKeague, I. W. (2013).
	\emph{Recovering gradients from sparsely observed functional
	  data.} Biometrics 69, 396-404 (2013).
	http://www.columbia.edu/~im2131/ps/growthrate-package-reference.pdf
}
